Move PXP-specific exceptions to existing errors package
