+++ backend/src/main/java/gov/cdc/usds/simplereport/api/model/errors/PatientLinkExpiredException.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.api.exceptions;
+package gov.cdc.usds.simplereport.api.model.errors;
 
 public class PatientLinkExpiredException extends RuntimeException {
   private static final long serialVersionUID = 1L;
