+++ backend/src/main/java/gov/cdc/usds/simplereport/service/TestOrderService.java
@@ -148,7 +148,7 @@ public class TestOrderService {
   }
 
   @AuthorizationConfiguration.RequirePermissionUpdateTest
-  public void updateTimeOfTestQuestions(String patientId, String pregnancy, Map<String, Boolean> symptoms,
+  public void updateTimeOfTestQuestions(UUID patientId, String pregnancy, Map<String, Boolean> symptoms,
       Boolean firstTest, LocalDate priorTestDate, String priorTestType, TestResult priorTestResult,
       LocalDate symptomOnsetDate, Boolean noSymptoms) {
     TestOrder order = retrieveTestOrder(patientId);
